check-sr-results: check Devicetree
We use dtc to check the captured Devicetree blob. For that, we add a new
`--dtc' command line option to allow specifying the dtc command line, and a
new `devicetree' property to the configuration file and schema.

Then we use dt-parser.py to check the dtc log. For that, we add a new
`--dt-parser' command line option to allow specifying the dt-parser command
line.

Signed-off-by: Vincent Stehlé <vincent.stehle@arm.com>