check-sr-results: re-create result.md
Enhance check-sr-results.py to attempt re-creating the result.md when it is
missing using the SCT parser.

We add an `sct-parser-result-md' property to the configuration file and we
add a `--parser' command line option to specify the SCT parser command.

The unit test copes with both cases of parser.py being in the $PATH or not.

Signed-off-by: Vincent Stehlé <vincent.stehle@arm.com>