capsule-tool: use lambda functions
Change all error and debug sanity check entries to lambda functions, which
receive the capsule as parameter.
No functional change in this commit, but this will allow printing not only
fixed error and debug messages, but also variable fields of the capsule.
While at it, document a bit the fields and interfaces.

Signed-off-by: Vincent Stehlé <vincent.stehle@arm.com>