dt-parser: add compatibles
We add a `--compatibles' command line option to supply a list of compatible
strings in a file, one by line.
We use that list to create a new `in_compatibles' property for the warnings
about missing schemas coming from dt-validate.
Then we use that new property in the rules to refine how we classify the
warnings:

- Warnings with no matching compatible string in the list remain classified
  as 'error no schema'.

- Warnings with a matching compatible string in the list are classified to
  less severe 'dt-validate warning no yaml schema'.

Also, update the schemas and the unit test.

Signed-off-by: Vincent Stehlé <vincent.stehle@arm.com>