check-sr-results: refine linux cache
Refine a bit our Linux cache directory structure and add a stamp, which we
can test to determine is caching is complete.
This should give us an "all or nothing" type of cache, which should be more
robust against e.g. download failures. It should also cope better with
future additions such as extracting the "compatibles" strings.

Signed-off-by: Vincent Stehlé <vincent.stehle@arm.com>