check-sr-results: check esp in more uefi logs
Record the ESPs we see during UEFI Shell sniff test log analysis for
deferred checks.

Add a new `must-have-esp' property and use it to mark those logs we want to
verify. For those files, we record all the device paths we see for deferred
checks. Also, update the schema and the documentation accordingly.

Add a deferred check to verify the recorded device paths against the ESPs
found.

While at it, add some more strings to check in the configuration files,
related with ESP.

Signed-off-by: Vincent Stehlé <vincent.stehle@arm.com>