check-sr-results: deferred checks
Add a deferred checks mechanism, which runs after checking all files and
dirs. This allows to decouple checks from the configuration file order.

Use the deferred checks for capsule GUIDs and ESRT. This was not necessary
due to the config file order in practice but is a good first user of the
new mechanism.

No functional change.

Signed-off-by: Vincent Stehlé <vincent.stehle@arm.com>