Limit line cleanup iterations
Limit the number of iterations to cleanup a line in `logreader.py'. This
greatly helps with some OS install logs, where the distro checks its
install medium and emits a ton of progress updates for example.

Add command line options to change this limit in `check-sr-results.py' and
`format-sr-results.py'.

Signed-off-by: Vincent Stehlé <vincent.stehle@arm.com>