ethernet-parser.py: fix corner case
When the number of expected ethernet interfaces is greater than the number
actually found, the `result' variable is used without being defined, which
leads to a crash. Make sure to define it in all cases.
Also, Add a corresponding unit test.

While at it, fix a log message and tune ethernet parser unit test printing.

Signed-off-by: Vincent Stehlé <vincent.stehle@arm.com>