guid.py: fix bug with leading zeroes
When the last 6 bytes `Node' field of a GUID has leading zeroes, they are
erroneously converted to spaces by the Guid.__str__() method; fix this.

This could cause havocs in e.g. `check-sr-results.py' when calling
`guid-tool.py' with a GUID as argument, as the GUID becomes split into two
arguments due to the spaces, ending up in a confusing error message.

Signed-off-by: Vincent Stehlé <vincent.stehle@arm.com>