guid.py: allow to deal with invalid GUIDs
When a new Guid object is created or when the date of a version 1 GUID is
retrieved, the validity of the GUID is verified and in the invalid cases,
an exception is raised.
While this is useful in the context of verification, this prevents
debugging invalid GUIDs further.

Add a `validate' parameter to the __init__() and get_datetime() methods, to
allow bypassing validation and preventing exceptions.
Those new parameters default to True to preserve previous behaviour.

We add a few doctests demonstrating the new possibilities.

Signed-off-by: Vincent Stehlé <vincent.stehle@arm.com>