dt-parser: ignore dt-validate missing schema warnings for u-boot nodes
U-Boot does publish a few Devicetree nodes to the OS with compatibles in
"u-boot,*". For example, looking at the Devicetrees in U-Boot v2025.10-rc4:

  u-boot,bootcount-rtc
  u-boot,bootcount-spi-flash
  u-boot,fs-loader
  u-boot,fwu-mdata-gpt
  u-boot,fwu-mdata-mtd
  u-boot,sysinfo-smbios

As they are clearly isolated in their own compatible "namespace" and do not
harm OS boot, add a rule to ignore those errors.

While at it, add a unit test.

Tested-by: Weizhao Ouyang <weizhao.ouyang@arm.com>
Signed-off-by: Vincent Stehlé <vincent.stehle@arm.com>